requirements.txt
# Daily Digest API - Python Dependencies

# FastAPI and web server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Configuration
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# AI/LLM
anthropic==0.18.1

# HTTP clients
httpx==0.26.0
requests==2.31.0

# Google APIs (Gmail, Calendar, Maps)
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0

# Async support
aiohttp==3.9.1

# Todoist API
todoist-api-python==2.1.4

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Optional: Redis for caching
redis==5.0.1

# Optional: Database
# sqlalchemy==2.0.25
# alembic==1.13.1

# Development dependencies (optional)
# pytest==7.4.4
# pytest-asyncio==0.23.3
# black==23.12.1
# flake8==7.0.0
